backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
openpyxl==3.1.5
httpx==0.27.2
